+++ packages/sdks/python/src/latitude_sdk/sdk/__init__.py
@@ -4,4 +4,6 @@ from .latitude import *
 from .logs import *
 from .projects import *
 from .prompts import *
+from .runs import *
 from .types import *
+from .versions import *